rename: FE폴더구조 변경
