+++ client/src/components/FriendsBar/FriendsBar.style.ts
