+++ client/src/components/MainContainer/Log.style.ts
