+++ client/src/components/MainContainer/MainContainer.style.ts
