+++ client/src/pages/WelcomePage.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoginMenu from '../components/WelcomePage/LoginMenu';
+import LoginMenu from '../components/LoginMenu/LoginMenu';
 
 const WelcomePage = () => {
   return <LoginMenu />;
